generated/xml_validation/xsd-validation__t3.html

<!DOCTYPE html>
<html lang="en">
  <head>
    <meta charset="utf-8" />
    <title>XSD Validation</title>
    <link rel="stylesheet" href="../lib/style.css" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <div id="content-wrapper">
      <div id="video-wrapper">
        <video></video>
        <div><div id="qr-code"></div></div>
        <div><div id="status-text"></div></div>
      </div>
      <div>To show/hide debug overlay press up/down or button beneath</div>
      <div id="debug-wrapper">
        <div>
          <div class="button" id="debug-button">Debug Overlay</div>
        </div>
        <div id="debug"></div>
        <div id="log"></div>
      </div>
    </div>
    <div id="info-overlay"></div>
    <script src="../lib/mozilla/object-keys-polyfill.js"></script>
    <script src="../lib/stefanpenner/es6-promise.min.js"></script>
    <script src="../lib/player.js"></script>
    <script src="../lib/manifest_parser.js"></script>
    <script src="../lib/mpd-parser.js"></script>
    <script src="../lib/wave-service.js"></script>
    <script src="../lib/davidshimjs/qrcode.js"></script>
    <script src="../lib/dpctf-testharness.js"></script>
    <script src="../lib/hbbtv.js"></script>
    <script src="../lib/xsdValidation/output.js"></script>
    <script>
      console.log("Execution begin of last script");
      // Global variables
      var TEST_INFO = {
        id: "xsd-validation",
        title: "XSD Validation for MPD files",
        description:
          "Template for XSD Validations.",
        code: "xsd-validation.html",
      };

      var video = document.querySelector("video");
      var qrCode = document.getElementById("qr-code");
      var statusText = document.getElementById("status-text");

      var videoContentModel = ["https://dash.akamaized.net/WAVE/vectors/cfhd_sets/15_30_60/t3/2023-09-01/stream.mpd"];
      var audioContentModel = [];

      var dpctfTest = new DpctfTest({
        testInfo: TEST_INFO,
        videoContentModel: videoContentModel,
        audioContentModel: audioContentModel,
        videoElement: video,
        qrCodeElement: qrCode,
        statusTextElement: statusText,
        infoOverlayElement: document.getElementById("info-overlay"),
        executeTest: executeTest,
        usePlayout: false,
      });

      var xml;
      var schemaStr;
      var schemaStr2;
      var schemaStr3;
      getXml();
      getXsd();
      getXlink();
      getXsd3();

      function executeTest(player, done, parameters) {
        console.log("In function executeTest");
        var minBufferDuration = parameters.minBufferDuration / 1000;
        player.startBuffering();

        dpctfTest.asyncTest(function (test) {
          console.log("In anonym function in asyncTest");

          console.log(parse(schemaStr));

          var query = location.search.replace(/\?/, "");
          var match = query.match(/redirect_time=([^&]+)/);
          var REDIRECT_TIME = match ? match[1] : null;
          if (!REDIRECT_TIME) REDIRECT_TIME = 5;
          console.log("Befor validate call");
          let result = validate();

          console.log(result);

          assert_true(result);
          test.done();

        }, "MPD Schema Validation");

        player.playOnBufferLoaded(minBufferDuration).then(done);
      }

      // save MPD file as utf-8 string
      async function getXml() {
        console.log("In function getXml");
        xml = await fetch(videoContentModel[0])
        .then(res => res.text());
        xml = JSON.parse( JSON.stringify( xml ) );
        if (!xml.includes('xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"') || !xml.includes('xsi:schemaLocation="urn:mpeg:DASH:schema:MPD:2011 DASH-MPD.xsd"')) {
            xml = xml.replace('<MPD', '<MPD xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="urn:mpeg:DASH:schema:MPD:2011 DASH-MPD.xsd"');
        }
      }
    
      // load main schema for validation
      async function getXsd() {
        console.log("In function getXsd");
        schemaStr = await fetch("../lib/xsdValidation/DASH-MPD.xsd")
        .then((res) => res.text());
        schemaStr = JSON.parse( JSON.stringify( schemaStr ) );
      }

      // load first dependencies of the main schema (xlink)
      async function getXlink() {
        console.log("In function getXlink");
        schemaStr2 = await fetch("../lib/xsdValidation/xlink.xsd")
        .then((res) => res.text());
        schemaStr2 = JSON.parse( JSON.stringify( schemaStr2 ) );
      }

      // load dependency of xlink (xml)
      async function getXsd3() {
        console.log("In function getXsd3");
        schemaStr3 = await fetch("../lib/xsdValidation/xml.xsd")
        .then((res) => res.text());
        schemaStr3 = JSON.parse( JSON.stringify( schemaStr3 ) );
      }

      function validate() {
        console.log("In function validate");
        /* console.log(xml);
        console.log(schemaStr);
        console.log(schemaStr2);
        console.log(schemaStr3); */
        // 1. write XSD-files into a virtual directory
        Module.FS.writeFile('/DASH-MPD.xsd', schemaStr);    // Hauptschema
        Module.FS.writeFile('/xlink.xsd', schemaStr2);  // Abhängigkeit
        Module.FS.writeFile('/xml.xsd', schemaStr3);      

        // 2. write the XML-string to memory
        const ptrXml = Module._malloc(lengthBytesUTF8(xml) + 1);
        stringToUTF8(xml, Module.HEAPU8, ptrXml, lengthBytesUTF8(xml) + 1);

        // 3. parse XML-document
        const xmlDoc = Module.ccall(
          'xmlReadMemory',
          'number',
          ['number', 'number', 'number', 'number', 'number'],
          [ptrXml, lengthBytesUTF8(xml), null, null, 0]
        );

        // 4. load scheme from file DASH-MPD.xsd (that imports work)
        const schemaParserCtx = Module.ccall(
          'xmlSchemaNewParserCtxt',
          'number',
          ['string'],
          ['/DASH-MPD.xsd']
        );

        // 5. parse schema and create validation context
        const schema = Module.ccall('xmlSchemaParse', 'number', ['number'], [schemaParserCtx]);
        const validCtx = Module.ccall('xmlSchemaNewValidCtxt', 'number', ['number'], [schema]);

        // 6. validate XML
        const result = Module.ccall(
          'xmlSchemaValidateDoc',
          'number',
          ['number', 'number'],
          [validCtx, xmlDoc]
        );

        // 7. release memory
        Module.ccall('xmlFreeDoc', 'void', ['number'], [xmlDoc]);
        Module.ccall('xmlSchemaFreeValidCtxt', 'void', ['number'], [validCtx]);
        Module.ccall('xmlSchemaFree', 'void', ['number'], [schema]);
        Module._free(ptrXml);

        console.log("result ", result);

        // 8. return result
        if (result === 0 ){
          return true;
        } else {
          return false;
        }
      }

      function stringToUTF8(str, heap, offset, maxBytesToWrite) {
        console.log("In function stringToUTF8");
        if (!(maxBytesToWrite > 0)) return 0;
        const utf8encoder = new TextEncoder();
        const encoded = utf8encoder.encode(str);
        const len = Math.min(encoded.length, maxBytesToWrite - 1);
        heap.set(encoded.subarray(0, len), offset);
        heap[offset + len] = 0;
        return len;
      }
    </script>
  </body>
</html>
